Added requiredFields to function creation
