--- src/lib/GridFunctions/gridfunc.test.js
@@ -1,5 +1,0 @@
-/*
-Test aggrid functions
-
-
-*/
