--- src/lib/GridParser/InitialParams/README
@@ -1,3 +1,0 @@
-pulls data
-sets initial sort, order_by values
-creates initial global object for calculations
