+++ src/lib/GridParser/README.md
@@ -1,0 +1,3 @@
+Main parser for json to page layout
+
+converts json to grid/header etc.
